+++ src/main/java/Ui.java
@@ -1,0 +1,7 @@
+public class Ui {
+
+    public void displayGreeting() {
+        System.out.println("Yahallo! I'm Duke\n"
+                + "What can I do for you?\n"); // print greeting message
+    }
+}
